Update views.py
what we are writing here is we want to tell the view that our host will be Redis for our counter to work properly, if the method is a post of data on the website then create a new item with the text that the end user has produced and return it with a redirect.  we will have items be equal to all items that way we can display all our text inputs on the homepage,  then we define our counter by using redis.incr which will display a number of times a end user has visited the page.